requirements.txt
asgiref==3.4.1
coverage==5.5
Django==3.2.7
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
flake8==3.9.2
flake8-html==0.4.1
importlib-metadata==4.8.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
psycopg2==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.10.0
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.2
zipp==3.5.0
